Добавлен вид после сдачи книги.

+++ src/main/java/library/controller/BorrowedBookController.java
@@ -1,10 +1,10 @@
 package library.controller;
 
-import library.FinallyPrice;
-import library.Librarian;
+import library.logic.AddUserModel;
+import library.logic.FinallyPrice;
+import library.logic.Librarian;
 import library.model.Book;
 import library.model.BorrowedBook;
-import library.model.User;
 import library.service.BookService;
 import library.service.BorrowedBookService;
 import library.service.UserService;
@@ -32,7 +32,9 @@ public class BorrowedBookController {
 
     private int discountRental;
 
-    private UserService userService;
+    private final UserService userService;
+
+    private final static int[] MONTHS_NUMBER = new int[]{1, 2, 3, 4, 5, 6};
 
     public BorrowedBookController(BookService bookService, BorrowedBookService borrowedBookService, UserService userService) {
         this.bookService = bookService;
@@ -42,34 +44,38 @@ public class BorrowedBookController {
 
     @GetMapping("/{id}")
     public String getCreationPage(Model model, @PathVariable int id, HttpSession session) {
-        checkInMenu(model, session);
+        AddUserModel.checkInMenu(model, session);
         Optional<Book> optionalBook = bookService.findById(id);
         if (optionalBook.isEmpty()) {
             model.addAttribute("message", "Книга не найдена");
             return "errors/404";
         }
-        model.addAttribute("deposit", "В качестве залога нужно внести " + optionalBook.get().getDepositPrice() + " рублей");
+        model.addAttribute("deposit",  optionalBook.get().getDepositPrice() + " рублей");
         model.addAttribute("book", optionalBook.get());
-        model.addAttribute("monthNum", new int[]{1, 2, 3, 4, 5, 6});
+        model.addAttribute("monthNum", MONTHS_NUMBER);
         return "borrowedBooks/create";
     }
 
     @PostMapping("/execution")
     public String execution(Model model, @ModelAttribute BorrowedBook borrowedBook, HttpSession session) {
-        checkInMenu(model, session);
+        AddUserModel.checkInMenu(model, session);
         saveBorrowedBook = borrowedBook;
-        Book book = bookService.findById(borrowedBook.getBookId()).get();
-        discountRental = FinallyPrice.getFinallyPrice(saveBorrowedBook.getTerm(), book.getRentalPrice());
-        if (borrowedBook.getDeposit() > book.getDepositPrice()) {
+        Optional<Book> optionalBook = bookService.findById(borrowedBook.getBookId());
+        if (optionalBook.isEmpty()) {
+            model.addAttribute("message", "Книга не найдена");
+            return "errors/404";
+        }
+        discountRental = FinallyPrice.getFinallyPrice(saveBorrowedBook.getTerm(), optionalBook.get().getRentalPrice());
+        if (borrowedBook.getDeposit() > optionalBook.get().getDepositPrice()) {
             model.addAttribute("message", "Сумма превосходит требуемую, попробуйте ещё раз");
             return "/errors/404";
         }
-        if (borrowedBook.getDeposit() < book.getDepositPrice()) {
+        if (borrowedBook.getDeposit() < optionalBook.get().getDepositPrice()) {
             model.addAttribute("message", "Внесенной суммы недостаточно, попробуйте ещё раз");
             return "/errors/404";
         }
         model.addAttribute("borrowedBook", borrowedBook);
-        model.addAttribute("bookMessage", book.getName());
+        model.addAttribute("bookMessage", optionalBook.get().getName());
         model.addAttribute("termMessage", borrowedBook.getTerm() + " месяц(ев)");
         model.addAttribute("depositMessage", borrowedBook.getDeposit() + " рублей");
         model.addAttribute("priceMessage", discountRental + " рублей");
@@ -79,7 +85,7 @@ public class BorrowedBookController {
 
     @PostMapping("/pay")
     public String save(Model model, @ModelAttribute BorrowedBook borrowedBook, HttpSession session) {
-        checkInMenu(model, session);
+        AddUserModel.checkInMenu(model, session);
         saveBorrowedBook.setRental(borrowedBook.getRental());
         model.addAttribute("borrowedBook", saveBorrowedBook);
         if (saveBorrowedBook.getRental() > discountRental) {
@@ -100,13 +106,19 @@ public class BorrowedBookController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(Model model, @PathVariable int id) {
+    public String delete(Model model, @PathVariable int id, HttpSession session) {
+        AddUserModel.checkInMenu(model, session);
+        Optional<BorrowedBook> optionalBorrowedBook = borrowedBookService.findById(id);
+        if (optionalBorrowedBook.isEmpty()) {
+            model.addAttribute("message", "Книга не найдена");
+            return "errors/404";
+        }
+        int deposit = optionalBorrowedBook.get().getDeposit();
         if (!borrowedBookService.deleteById(id)) {
             model.addAttribute("message", "Книга с указанным идентификатором не найдена");
         }
-        model.addAttribute("borrowedBooks", borrowedBookService.findAll());
-        model.addAttribute("books", bookService.findAll());
-        return "personalAccount/personalAccount";
+        model.addAttribute("deposit", deposit + " рублей.");
+        return "/librarian/successfully";
     }
 
     @GetMapping("/download/{id}")
@@ -117,27 +129,25 @@ public class BorrowedBookController {
             return;
         }
         BorrowedBook borrowedBook = optionalBorrowedBook.get();
+        Optional<Book> optionalBook = bookService.findById(borrowedBook.getBookId());
+        if (optionalBook.isEmpty()) {
+            model.addAttribute("message", "Книга не найдена");
+            return;
+        }
         Book book = bookService.findById(optionalBorrowedBook.get().getBookId()).get();
         String userName = userService.findUserNameById(borrowedBook.getUserId());
-        response.setContentType("application/octet-stream");
-            String headerKey = "Content-Disposition";
-            String headerValue = "attachment; filename = " + borrowedBook.getId();
-            response.setHeader(headerKey, headerValue);
-            try (ServletOutputStream outputStream = response.getOutputStream()) {
-                Librarian librarian = new Librarian();
-                byte[] content = librarian.receipt(borrowedBook, book, userName).getBytes();
-                outputStream.write(content);
-
-            }
-    }
-
-    private void checkInMenu(Model model, HttpSession session) {
-        User user = (User) session.getAttribute("user");
-        if (user == null) {
-            user = new User();
-            user.setName("Гость");
+        if (userName.equals("Гость")) {
+            model.addAttribute("message", "Пользователь не найден");
         }
-        model.addAttribute("user", user);
+        response.setContentType("application/octet-stream");
+        String headerKey = "Content-Disposition";
+        String headerValue = "attachment; filename = " + borrowedBook.getId();
+        response.setHeader(headerKey, headerValue);
+        try (ServletOutputStream outputStream = response.getOutputStream()) {
+            Librarian librarian = new Librarian();
+            byte[] content = librarian.receipt(borrowedBook, book, userName).getBytes();
+            outputStream.write(content);
+        }
     }
 
 }
